### 2018-08-15 #894 @whedon
[PRE REVIEW]: compboost: Modular Framework for Component-Wise Boosting
**Submitting author:** @schalkdaniel (<a href="http://orcid.org/0000-0003-0950-1947">Daniel Schalk</a>)
**Repository:** <a href="https://github.com/schalkdaniel/compboost" target ="_blank">https://github.com/schalkdaniel/compboost</a>
**Version:** v0.1.0
