### 2018-08-15 f109dc7
remove unused reference in paper.bib
