+++ paper/paper.bib
@@ -57,14 +57,3 @@
   publisher={Springer},
   doi={10.1007/978-1-4614-6868-4}
 }
-
-@article{,
-  title = {RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
-  author = {Dirk Eddelbuettel and Conrad Sanderson},
-  journal = {Computational    Statistics and Data Analysis},
-  year = {2014},
-  volume = {71},
-  month = {March},
-  pages = {1054--1063},
-  doi = {10.1016/j.csda.2013.02.005}
-}
