### 2018-08-29 #894 @schalkdaniel
Hi, the editor assignment is now pending for 2 weeks. Is there anything I can do to speed up the process?
