### 2018-08-29 #894 @arfon
> Hi, the editor assignment is now pending for 2 weeks. Is there anything I can do to speed up the process?

Apologies. We're very busy right now. I'll see if I can encourage some of our editors to get this moving.
