+++ R/zzz.R
@@ -2,4 +2,5 @@
 NULL
 
 #' @import Rcpp
+#' @import RcppArmadillo
 NULL
