+++ NAMESPACE
@@ -23,4 +23,5 @@ export(boostLinear)
 export(boostSplines)
 export(getCustomCppExample)
 import(Rcpp)
+import(RcppArmadillo)
 useDynLib(compboost, .registration = TRUE)

[…]
