### 2018-08-29 87436f0
add reverse dependencies check
