+++ revdep/README.md
@@ -1,0 +1,40 @@
+# Setup
+
+## Platform
+
+|setting  |value                        |
+|:--------|:----------------------------|
+|version  |R version 3.5.1 (2018-07-02) |
+|system   |x86_64, linux-gnu            |
+|ui       |X11                          |
+|language |(EN)                         |
+|collate  |en_US.UTF-8                  |
+|tz       |Europe/Berlin                |
+|date     |2018-08-29                   |
+
+## Packages
+
+|package       |*  |version     |date       |source                            |
+|:-------------|:--|:-----------|:----------|:---------------------------------|
+|checkmate     |   |1.8.5       |2017-10-24 |cran (@1.8.5)                     |
+|compboost     |   |0.1.0       |2018-08-29 |local (schalkdaniel/compboost@NA) |
+|ggplot2       |   |3.0.0       |2018-07-03 |cran (@3.0.0)                     |
+|glue          |   |1.3.0       |2018-07-17 |cran (@1.3.0)                     |
+|gridExtra     |   |2.3         |2017-09-09 |cran (@2.3)                       |
+|knitr         |   |1.20        |2018-02-20 |cran (@1.20)                      |
+|mboost        |   |2.9-1       |2018-08-22 |cran (@2.9-1)                     |
+|mlr           |   |2.13        |2018-08-28 |cran (@2.13)                      |
+|R6            |   |2.2.2       |2017-06-17 |cran (@2.2.2)                     |
+|Rcpp          |   |0.12.18     |2018-07-23 |cran (@0.12.18)                   |
+|RcppArmadillo |   |0.9.100.5.0 |2018-08-16 |cran (@0.9.100)                   |
+|rmarkdown     |   |1.10        |2018-06-11 |cran (@1.10)                      |
+|testthat      |   |2.0.0       |2017-12-13 |cran (@2.0.0)                     |
+|titanic       |   |0.1.0       |2015-08-31 |cran (@0.1.0)                     |
+
+# Check results
+
+0 packages
+
+
+
+
