+++ Readme.Rmd
@@ -2,6 +2,8 @@
 output: github_document
 ---
 
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
 ```{r, include=FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE
