+++ Readme.Rmd
@@ -1,132 +1,0 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-```{r, include=FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE
-  , fig.path = "Readme_files/"
-)
-```
-
-<img align="right" src="docs/images/cboost_hexagon.png" width="100px">
-
-[![Build Status](https://travis-ci.org/schalkdaniel/compboost.svg?branch=master)](https://travis-ci.org/schalkdaniel/compboost) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/schalkdaniel/compboost?branch=master&svg=true)](https://ci.appveyor.com/project/schalkdaniel/compboost) [![Coverage Status](https://coveralls.io/repos/github/schalkdaniel/compboost/badge.svg?branch=master)](https://coveralls.io/github/schalkdaniel/compboost?branch=master)
-
-## compboost: Fast and Flexible Component-Wise Boosting Framework
-
-Component-wise boosting applies the boosting framework to
-statistical models, e.g., general additive models using component-wise smoothing 
-splines. Boosting these kinds of models maintains interpretability and enables 
-unbiased model selection in high dimensional feature spaces. 
-
-The `R` package `compboost` is an alternative implementation of component-wise
-boosting written in `C++` to obtain high runtime
-performance and full memory control. The main idea is to provide a modular
-class system which can be extended without editing the
-source code. Therefore, it is possible to use `R` functions as well as
-`C++` functions for custom base-learners, losses, logging mechanisms or 
-stopping criteria.
-
-For an introduction and overview about the functionality visit the [project page](https://schalkdaniel.github.io/compboost/).
-
-## Installation
-
-#### Developer version:
-
-```r
-devtools::install_github("schalkdaniel/compboost")
-```
-
-## Examples
-
-This examples are rendered using <code>compboost `r packageVersion("compboost")`</code>.
-
-To be as flexible as possible one should use the `R6` API do define base-learner, losses, stopping criteria, or optimizer as desired. Another option is to use wrapper functions as described on the [project page](https://schalkdaniel.github.io/compboost/).
-
-```{r cboost, fig.show="hide", out.width="70%"}
-library(compboost)
-
-# Check installed version:
-packageVersion("compboost")
-
-# Load data set with binary classification task:
-data(PimaIndiansDiabetes, package = "mlbench")
-# Create categorical feature:
-PimaIndiansDiabetes$pregnant.cat = ifelse(PimaIndiansDiabetes$pregnant == 0, "no", "yes")
-
-# Define Compboost object:
-cboost = Compboost$new(data = PimaIndiansDiabetes, target = "diabetes", loss = LossBinomial$new())
-cboost
-
-# Add p-spline base-learner with default parameter:
-cboost$addBaselearner(feature = "pressure", id = "spline", bl.factory = BaselearnerPSpline)
-
-# Add another p-spline learner with custom parameters:
-cboost$addBaselearner(feature = "age", id = "spline", bl.factory = BaselearnerPSpline, degree = 3, 
-  knots = 10, penalty = 4, differences = 2)
-
-# Add categorical feature (as single linear base-learner):
-cboost$addBaselearner(feature = "pregnant.cat", id = "category", bl.factory = BaselearnerPolynomial,
-	degree = 1, intercept = FALSE)
-
-# Check all registered base-learner:
-cboost$getBaselearnerNames()
-
-# Train model:
[…]
