+++ Readme.md
@@ -1,182 +1,0 @@
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-<img align="right" src="docs/images/cboost_hexagon.png" width="100px">
-
-[![Build
-Status](https://travis-ci.org/schalkdaniel/compboost.svg?branch=master)](https://travis-ci.org/schalkdaniel/compboost)
-[![AppVeyor Build
-Status](https://ci.appveyor.com/api/projects/status/github/schalkdaniel/compboost?branch=master&svg=true)](https://ci.appveyor.com/project/schalkdaniel/compboost)
-[![Coverage
-Status](https://coveralls.io/repos/github/schalkdaniel/compboost/badge.svg?branch=master)](https://coveralls.io/github/schalkdaniel/compboost?branch=master)
-
-## compboost: Fast and Flexible Component-Wise Boosting Framework
-
-Component-wise boosting applies the boosting framework to statistical
-models, e.g., general additive models using component-wise smoothing
-splines. Boosting these kinds of models maintains interpretability and
-enables unbiased model selection in high dimensional feature spaces.
-
-The `R` package `compboost` is an alternative implementation of
-component-wise boosting written in `C++` to obtain high runtime
-performance and full memory control. The main idea is to provide a
-modular class system which can be extended without editing the source
-code. Therefore, it is possible to use `R` functions as well as `C++`
-functions for custom base-learners, losses, logging mechanisms or
-stopping criteria.
-
-For an introduction and overview about the functionality visit the
-[project page](https://schalkdaniel.github.io/compboost/).
-
-## Installation
-
-#### Developer version:
-
-``` r
-devtools::install_github("schalkdaniel/compboost")
-```
-
-## Examples
-
-This examples are rendered using <code>compboost 0.1.0</code>.
-
-To be as flexible as possible one should use the `R6` API do define
-base-learner, losses, stopping criteria, or optimizer as desired.
-Another option is to use wrapper functions as described on the [project
-page](https://schalkdaniel.github.io/compboost/).
-
-``` r
-library(compboost)
-
-# Check installed version:
-packageVersion("compboost")
-## [1] '0.1.0'
-
-# Load data set with binary classification task:
-data(PimaIndiansDiabetes, package = "mlbench")
-# Create categorical feature:
-PimaIndiansDiabetes$pregnant.cat = ifelse(PimaIndiansDiabetes$pregnant == 0, "no", "yes")
-
-# Define Compboost object:
-cboost = Compboost$new(data = PimaIndiansDiabetes, target = "diabetes", loss = LossBinomial$new())
-cboost
-## Component-Wise Gradient Boosting
-## 
-## Trained on PimaIndiansDiabetes with target diabetes
-## Number of base-learners: 0
-## Learning rate: 0.05
-## Iterations: 0
-## Positive class: neg
-## 
-## LossBinomial Loss:
-## 
-##   Loss function: L(y,x) = log(1 + exp(-2yf(x))
-## 
-## 
-
-# Add p-spline base-learner with default parameter:
-cboost$addBaselearner(feature = "pressure", id = "spline", bl.factory = BaselearnerPSpline)
-
-# Add another p-spline learner with custom parameters:
-cboost$addBaselearner(feature = "age", id = "spline", bl.factory = BaselearnerPSpline, degree = 3, 
-  knots = 10, penalty = 4, differences = 2)
-## Warning in .handleRcpp_BaselearnerPSpline(degree = 3, knots = 10, penalty =
[…]
