### 2018-08-29 4efef62
fix miss spellings
