+++ R/boost_splines.R
@@ -6,7 +6,7 @@
 #' initializing the model \code{boostSpline} also fits as many iterations as given 
 #' by the user through \code{iters}. 
 #' 
-#' The returend object is an object of the \code{Compboost} class which then can be 
+#' The returned object is an object of the \code{Compboost} class which then can be 
 #' used for further analyses (see \code{?Compboost} for details). 
 #'
 #' @return Usually a model of class \code{Compboost}. This model is an \code{R6} object
@@ -29,10 +29,10 @@
 #'   Number of iterations that are trained.
 #' @param trace [\code{integer(1)}]\cr
 #'   Integer indicating how often a trace should be printed. Specifying \code{trace = 10}, then every
-#'   10th iteration is printet. If no trace should be printed set \code{trace = 0}. Default is
+#'   10th iteration is printed. If no trace should be printed set \code{trace = 0}. Default is
 #'   -1 which means that we set \code{trace} at a value that 40 iterations are printed.
 #' @param degree [\code{integer(1)}]\cr
-#'   Polynomial degree of the splines used for modelling. Note that the number of parameter
+#'   Polynomial degree of the splines used for modeling. Note that the number of parameter
 #'   increases with the degrees.
 #' @param n.knots [\code{integer(1)}]\cr
 #'   Number of equidistant "inner knots". The real number of used knots also depends on
@@ -42,13 +42,13 @@
 #'   The higher penalty, the higher the smoothness.
 #' @param differences [\code{integer(1)}]\cr
 #'   Number of differences that are used for penalization. The higher this value is, the
-#'   more function values of neighbour knots are forced to be more similar which results
-#'   in a smoother courve.
+#'   more function values of neighbor knots are forced to be more similar which results
+#'   in a smoother curve.
 #' @param data.source [\code{S4 Data}]\cr
-#'   Unitialized \code{S4 Data} object which is used to store the data. At the moment
+#'   Uninitialized \code{S4 Data} object which is used to store the data. At the moment
 #'   just in memory training is supported.
 #' @param data.target [\code{S4 Data}]\cr
-#'   Unitialized \code{S4 Data} object which is used to store the data. At the moment
+#'   Uninitialized \code{S4 Data} object which is used to store the data. At the moment
 #'   just in memory training is supported.
 #' @examples
 #' mod = boostSplines(data = iris, target = "Sepal.Length", loss = LossQuadratic$new())
