+++ R/get_custom_cpp_example.R
@@ -38,8 +38,8 @@
 #'   of a trained compboost object.
 #'
 #' @param example [\code{character(1)}] \cr
-#'   Character value indicating if an example for the baselearner or for the
-#'   loss should be returned. The values for esample has to be one of 
+#'   Character value indicating if an example for the base-learner or for the
+#'   loss should be returned. The values, for example, has to be one of 
 #'   \code{blearner} or \code{loss}.
 #' @param silent [\code{logical(1)}] \cr
 #'   Logical value indicating if the example code should be printed to the
