+++ src/compboost_modules.cpp
@@ -64,7 +64,7 @@ protected:
 //'
 //' \code{InMemoryData} creates an data object which can be used as source or
 //' target object within the base-learner factories of \code{compboost}. The
-//' convention to initialise target data is to call the constructor without
+//' convention to initialize target data is to call the constructor without
 //' any arguments.
 //'
 //' @format \code{\link{S4}} object.
@@ -203,7 +203,7 @@ RCPP_MODULE (data_module)
 // -------------------------------------------------------------------------- //
 
 // Abstract class. This one is given to the factory list. The factory list then
-// handles all factorys equally. It does not differ between a polynomial or
+// handles all factories equally. It does not differ between a polynomial or
 // custom factory:
 class BaselearnerFactoryWrapper
 {
@@ -219,7 +219,7 @@ protected:
 };
 
 
-//' Baselearner factory to make polynomial regression
+//' Base-learner factory to make polynomial regression
 //'
 //' \code{BaselearnerPolynomial} creates a polynomial base-learner factory
 //'  object which can be registered within a base-learner list and then used
@@ -268,7 +268,7 @@ protected:
 //' @section Methods:
 //' \describe{
 //' \item{\code{getData()}}{Get the data matrix of the target data which is used
-//'   for modelling.}
+//'   for modeling.}
 //' \item{\code{transformData(X)}}{Transform a data matrix as defined within the
 //'   factory. The argument has to be a matrix with one column.}
 //' \item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
@@ -383,7 +383,7 @@ public:
 //'   Degree of the spline functions to interpolate the knots.
 //' }
 //' \item{\code{n_knots} [\code{integer(1)}]}{
-//'   Number of \strong{inner knots}. To prefent weird behaviour on the edges
+//'   Number of \strong{inner knots}. To prevent weird behavior on the edges
 //'   the inner knots are expanded by \eqn{\mathrm{degree} - 1} additional knots.
 //' }
 //' \item{\code{penalty} [\code{numeric(1)}]}{
@@ -411,7 +411,7 @@ public:
 //' @section Methods:
 //' \describe{
 //' \item{\code{getData()}}{Get the data matrix of the target data which is used
-//'   for modelling.}
+//'   for modeling.}
 //' \item{\code{transformData(X)}}{Transform a data matrix as defined within the
 //'   factory. The argument has to be a matrix with one column.}
 //' \item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
@@ -521,7 +521,7 @@ public:
 //'   For details see the \code{Details}.
 //' }
 //' \item{\code{extractParameter} [\code{function}]}{
-//'   \code{R} function to extract the parameter of the object returend by
+//'   \code{R} function to extract the parameter of the object returned by
 //'   \code{train}. For details see the \code{Details}.
 //' }
 //' }
@@ -558,7 +558,7 @@ public:
 //' @section Methods:
 //' \describe{
 //' \item{\code{getData()}}{Get the data matrix of the target data which is used
-//'   for modelling.}
+//'   for modeling.}
 //' \item{\code{transformData(X)}}{Transform a data matrix as defined within the
 //'   factory. The argument has to be a matrix with one column.}
 //' \item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
@@ -687,7 +687,7 @@ public:
 //' @section Methods:
 //' \describe{
 //' \item{\code{getData()}}{Get the data matrix of the target data which is used
-//'   for modelling.}
+//'   for modeling.}
[…]
