+++ man/BaselearnerCustom.Rd
@@ -39,7 +39,7 @@ BaselearnerCustom$new(data_source, data_target, instantiateData, train,
   For details see the \code{Details}.
 }
 \item{\code{extractParameter} [\code{function}]}{
-  \code{R} function to extract the parameter of the object returend by
+  \code{R} function to extract the parameter of the object returned by
   \code{train}. For details see the \code{Details}.
 }
 }
@@ -82,7 +82,7 @@ BaselearnerCustom$new(data_source, data_target, instantiateData, train,
 
 \describe{
 \item{\code{getData()}}{Get the data matrix of the target data which is used
-  for modelling.}
+  for modeling.}
 \item{\code{transformData(X)}}{Transform a data matrix as defined within the
   factory. The argument has to be a matrix with one column.}
 \item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
