+++ man/LossBinomial.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/RcppExports.R
 \name{LossBinomial}
 \alias{LossBinomial}
-\title{0-1 Loss for binary classification derifed of the binomial distribution}
+\title{0-1 Loss for binary classification derived of the binomial distribution}
 \format{\code{\link{S4}} object.}
 \description{
 This loss can be used for binary classification. The coding we have chosen
