### 2018-08-29 eec8e5a
fix typos
