### 2018-08-30 c677af9
update tests which often crashes on windows
