+++ tests/testthat/test_compboost_internal.R
@@ -31,9 +31,9 @@ test_that("Compboost loggs correctly", {
   expect_silent({ loss.quadratic = LossQuadratic$new() })
   expect_silent({ optimizer = OptimizerCoordinateDescent$new() })
   expect_silent({ log.iterations = LoggerIteration$new(TRUE, iter.max) })
-  expect_silent({ log.time.ms    = LoggerTime$new(TRUE, 50000, "microseconds") })
-  expect_silent({ log.time.sec   = LoggerTime$new(TRUE, 2, "seconds") })
-  expect_silent({ log.time.min   = LoggerTime$new(TRUE, 1, "minutes") })
+  expect_silent({ log.time.ms    = LoggerTime$new(TRUE, 200000, "microseconds") })
+  expect_silent({ log.time.sec   = LoggerTime$new(TRUE, 10, "seconds") })
+  expect_silent({ log.time.min   = LoggerTime$new(TRUE, 10, "minutes") })
   expect_silent({ log.inbag      = LoggerInbagRisk$new(FALSE, loss.quadratic, 0.01) })
   expect_silent({ log.oob        = LoggerOobRisk$new(FALSE, loss.quadratic, 0.01, eval.oob.test, y) })
   expect_silent({ logger.list = LoggerList$new() })
