### 2018-08-30 602540a
Update cran-comments.md
