+++ cran-comments.md
@@ -1,6 +1,8 @@
 ## Test environments
 * local ubuntu 18.04, R 3.5.1
 * ubuntu 14.04 (on travis-ci), R 3.5.0
+* local Windows 10, R 3.5.1
+* Windows Server 2012 (on Appveyor), R 3.5.1
 * win-builder (devel and release)
 
 ## R CMD check results
