### 2018-08-30 db7f9a6
Update cran-comments.md
