### 2018-08-30 64da9bf
Update cran-comments.md
