### 2018-09-02 071c58a
Stupid Rmd -> md kompilation
