### 2018-09-04 #894 @arfon
:wave: @jakevdp - would you be able to edit this submission for JOSS?
