### 2018-09-11 7f251d4
move rcpp(armadillo) to suggests
