+++ DESCRIPTION
@@ -17,8 +17,6 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.1.0
 Imports:
-  Rcpp (>= 0.11.2),
-  RcppArmadillo (>= 0.8.500.0.0),
   methods,
   glue,
   R6,
@@ -27,6 +25,8 @@ LinkingTo:
   Rcpp,
   RcppArmadillo
 Suggests:
+  Rcpp (>= 0.11.2),
+  RcppArmadillo (>= 0.8.500.0.0),
   ggplot2,
   testthat,
   rpart,
