### 2018-09-11 01fea18
move RcppArmadillo to suggests
