### 2018-09-11 0d754b8
remove RcppArmadillo from NAMESPACE and zzz.R
