+++ NAMESPACE
@@ -23,5 +23,4 @@ export(boostLinear)
 export(boostSplines)
 export(getCustomCppExample)
 import(Rcpp)
-import(RcppArmadillo)
 useDynLib(compboost, .registration = TRUE)
