### 2018-09-20 #894 @arfon
Hi @schalkdaniel - could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission?
