### 2018-09-21 #967 @brainstorm
@moonso, please read the instructions on this issue, you are now the reviewer for this JOSS paper, let me know if anything is unclear about how to proceed with it :)
