### 2018-09-24 #967 @moonso
Hi @brainstorm ! Thanks for asking me to review, I will do my best here. To start with I have a few questions:

1. The theory and real world use of this project are a bit over my expertise. When I read the [reviewer guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html) it is mostly about testing the functionality of the package so I guess this is my focus?

1. When I post issues in the package should I use some special words to indicate that I am a reviewer? In this case the [LICENCE](https://github.com/schalkdaniel/compboost/blob/master/LICENSE) does not follow any of the approved ones as far as I can tell. Not sure how to communicate that in the issue.

Thx!
