### 2018-09-24 #967 @brainstorm
1. Yes, that's right @moonso, your overall task (TL;DR'd) here is to validate the operation/usefulness/completeness of this package.

2. I already assigned you as a reviewer, so whedon, our friendly bot knows you already, no worries ;)

@schalkdaniel, as @moonso points out, you should use a valid, [OSI approved LICENSE](https://opensource.org/licenses/alphabetical).

@moonso You can move on into reviewing the other points while the author addresses things. Thanks a lot for helping out JOSS! ;)
