+++ README.md
@@ -1,20 +1,35 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
 <img align="right" src="docs/images/cboost_hexagon.png" width="100px">
 
-[![Build Status](https://travis-ci.org/schalkdaniel/compboost.svg?branch=master)](https://travis-ci.org/schalkdaniel/compboost) [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/schalkdaniel/compboost?branch=master&svg=true)](https://ci.appveyor.com/project/schalkdaniel/compboost) [![Coverage Status](https://coveralls.io/repos/github/schalkdaniel/compboost/badge.svg?branch=master)](https://coveralls.io/github/schalkdaniel/compboost?branch=master) [![license](https://img.shields.io/github/license/mashape/apistatus.svg?maxAge=2592000)](#license)
+[![Build
+Status](https://travis-ci.org/schalkdaniel/compboost.svg?branch=master)](https://travis-ci.org/schalkdaniel/compboost)
+[![AppVeyor Build
+Status](https://ci.appveyor.com/api/projects/status/github/schalkdaniel/compboost?branch=master&svg=true)](https://ci.appveyor.com/project/schalkdaniel/compboost)
+[![Coverage
+Status](https://coveralls.io/repos/github/schalkdaniel/compboost/badge.svg?branch=master)](https://coveralls.io/github/schalkdaniel/compboost?branch=master)
+[![license](https://img.shields.io/github/license/mashape/apistatus.svg?maxAge=2592000)](#license)
 
-compboost: Fast and Flexible Component-Wise Boosting Framework
---------------------------------------------------------------
+## compboost: Fast and Flexible Component-Wise Boosting Framework
 
-Component-wise boosting applies the boosting framework to statistical models, e.g., general additive models using component-wise smoothing splines. Boosting these kinds of models maintains interpretability and enables unbiased model selection in high dimensional feature spaces.
+Component-wise boosting applies the boosting framework to statistical
+models, e.g., general additive models using component-wise smoothing
+splines. Boosting these kinds of models maintains interpretability and
+enables unbiased model selection in high dimensional feature spaces.
 
-The `R` package `compboost` is an alternative implementation of component-wise boosting written in `C++` to obtain high runtime performance and full memory control. The main idea is to provide a modular class system which can be extended without editing the source code. Therefore, it is possible to use `R` functions as well as `C++` functions for custom base-learners, losses, logging mechanisms or stopping criteria.
+The `R` package `compboost` is an alternative implementation of
+component-wise boosting written in `C++` to obtain high runtime
+performance and full memory control. The main idea is to provide a
+modular class system which can be extended without editing the source
+code. Therefore, it is possible to use `R` functions as well as `C++`
+functions for custom base-learners, losses, logging mechanisms or
+stopping criteria.
 
-For an introduction and overview about the functionality visit the [project page](https://schalkdaniel.github.io/compboost/).
+For an introduction and overview about the functionality visit the
+[project page](https://schalkdaniel.github.io/compboost/).
 
-Installation
-------------
+## Installation
 
 #### Developer version:
 
@@ -22,12 +37,14 @@ Installation
 devtools::install_github("schalkdaniel/compboost")
 ```
 
-Examples
---------
+## Examples
 
 This examples are rendered using <code>compboost 0.1.0</code>.
 
[…]
