### 2018-09-25 #967 @schalkdaniel
Hi @moonso @brainstorm, I have added another file `LICENSE.md` containing the plain text of the MIT license. The file `LICENSE` is required by CRAN and should just include the year and copyright holder (if I understood that correctly).
