### 2018-09-26 #967 @brainstorm
Weird, are you sure @schalkdaniel?:

https://cran.r-project.org/web/licenses/

According to:

https://cran.r-project.org/doc/contrib/Leisch-CreatingPackages.pdf

The format you mention seems to belong to `DESCRIPTION`?
