### 2018-09-26 #967 @schalkdaniel
But isn't that what I have done @brainstorm? Maybe I miss something, but my `DESCRIPTION` file contains
```
License: MIT + file LICENSE
```
and refers to the `LICENSE` file as specified in your link and also [here](http://r-pkgs.had.co.nz/description.html). Some other repositories follow the same guidelines, e.g. [iml](https://github.com/christophm/iml), [pdfsearch](https://github.com/lebebr01/pdfsearch), or [checkr](https://github.com/poissonconsulting/checkr).
