### 2018-09-26 7b1442b
upwrap expensive (>10 sec) examples with donttest
