+++ R/RcppExports.R
@@ -411,6 +411,7 @@ NULL
 #' \item{\code{summarizeFactory()}}{Summarize the base-learner factory object.}
 #' }
 #' @examples
+#' \donttest{
 #' # Sample data:
 #' data.mat = cbind(1, 1:10)
 #' y = 2 + 3 * 1:10
@@ -434,7 +435,7 @@ NULL
 #'
 #' # Transform data manually:
 #' custom.cpp.factory$transformData(data.mat)
-#'
+#' }
 #' @export BaselearnerCustomCpp
 NULL
 
@@ -773,14 +774,13 @@ NULL
 #'   \url{https://schalkdaniel.github.io/compboost/cpp_man/html/classloss_1_1_custom_cpp_loss.html}.
 #'
 #' @examples
-#'
+#' \donttest{
 #' # Load loss functions:
 #' Rcpp::sourceCpp(code = getCustomCppExample(example = "loss", silent = TRUE))
 #'
 #' # Create new custom quadratic loss:
 #' my.cpp.loss = LossCustomCpp$new(lossFunSetter(), gradFunSetter(), constInitFunSetter())
-#'
-#'
+#' }
 #' @export LossCustomCpp
 NULL
 
