+++ man/LossCustomCpp.Rd
@@ -41,12 +41,11 @@ LossCustomCpp$new(loss_ptr, grad_ptr, const_init_ptr)
 }
 
 \examples{
-
+\donttest{
 # Load loss functions:
 Rcpp::sourceCpp(code = getCustomCppExample(example = "loss", silent = TRUE))
 
 # Create new custom quadratic loss:
 my.cpp.loss = LossCustomCpp$new(lossFunSetter(), gradFunSetter(), constInitFunSetter())
-
-
+}
 }
