### 2018-09-27 #967 @schalkdaniel
Ah, yes I see. Originally there were one `LICENSE` and the plain text of the MIT license was included in the `Readme`. The intention to include another `LICENSE.md` file came after looking at repositories from the [tidyverse](https://github.com/tidyverse) (e.g. [dplyr](https://github.com/tidyverse/dplyr)) where they do it like this. :)
