### 2018-09-27 #967 @moonso
Hi @schalkdaniel @brainstorm . I'm done with the review now, nice job @schalkdaniel 😸 . The documentation is clear and comprehensive, examples could be followed.
I went through a little installation hell, nice to see that it exists in R to 😛 , it had nothing to do with this package though. It was to install 'devtools' on my Mac OS.

My only remark is the contribution of the co-authors. Thomas Janek has contributed a fairly small part of the code compare to the main author. Also it is unclear how Bernd Bischl have contributed, I guess he is the PI but I think it should be stated somewhere how this person have contributed.

Again good job!

Best

/Måns
