### 2018-09-28 #967 @brainstorm
@arfon Do we have any established policies for contributions that are not reflected/seen on github?

Other than that, this is a thumbs up submission, thanks everyone! 👍
