### 2018-09-28 #967 @arfon
> @arfon Do we have any established policies for contributions that are not reflected/seen on github?

Ultimately, authorship is left up to the authors but we have some guidelines here: https://joss.readthedocs.io/en/latest/review_criteria.html#authorship
