### 2018-10-01 #967 @brainstorm
Alright, nothing seems off as far as I can see regarding authorship then and the author stated publicly who contributed where.
