### 2018-10-01 4a23d59
Create CONTRIBUTORS.md
