+++ CONTRIBUTORS.md
@@ -1,0 +1,10 @@
+## Contributors of `compboost`
+
+Committers are people who have made substantial contribution to the project and granted write access to the project.
+
+-   [Daniel Schalk](https://github.com/schalkdaniel), University of Munich
+    -   Daniel is the creator and maintainer of the project.
+-   [Janek Thomas](https://github.com/ja-thomas), University of Munich
+    -   Janek gives feedback/guidance for the package design and structure and implemets the intial `R` API.
+-   [Bernd Bischl](https://github.com/berndbischl), University of Munich
+    -   Bernd gives feedback/guidance for the package design and structure and initiated the project.
