### 2018-10-01 #967 @schalkdaniel
@brainstorm I have added a file `CONTRIBUTORS.md` and linked on this file from the Readme.
