+++ binder/install.R
@@ -1,0 +1,4 @@
+install.packages("ggplot2")
+install.pacakges("ggthemes")
+install.packages("mlbench")
+remotes::isntall_github("schalkdaniel/compboost")
