+++ binder/tutorial_intro.Rmd
@@ -1,0 +1,79 @@
+## Load Package 
+
+```{r}
+library(compboost)
+
+# Check installed version:
+packageVersion("compboost")
+```
+
+## Access Dataset
+
+```{r}
+# Load data set with binary classification task:
+data(PimaIndiansDiabetes, package = "mlbench")
+# Create categorical feature:
+PimaIndiansDiabetes$pregnant.cat = ifelse(PimaIndiansDiabetes$pregnant == 0, "no", "yes")
+```
+
+## Initialize a Compboost Model
+
+```{r}
+# Define Compboost object:
+cboost = Compboost$new(data = PimaIndiansDiabetes, target = "diabetes", loss = LossBinomial$new())
+cboost
+```
+
+## Add Base-Learner
+
+```{r}
+# Add p-spline base-learner with default parameter:
+cboost$addBaselearner(feature = "pressure", id = "spline", bl.factory = BaselearnerPSpline)
+
+# Add another p-spline learner with custom parameters:
+cboost$addBaselearner(feature = "age", id = "spline", bl.factory = BaselearnerPSpline, degree = 3, 
+  n.knots = 10, penalty = 4, differences = 2)
+
+# Add categorical feature (as single linear base-learner):
+cboost$addBaselearner(feature = "pregnant.cat", id = "category", bl.factory = BaselearnerPolynomial,
+  degree = 1, intercept = FALSE)
+
+# Check all registered base-learner:
+cboost$getBaselearnerNames()
+```
+
+## Initial Training 
+
+```{r}
+# Train model:
+cboost$train(1000L, trace = 200L)
+cboost
+```
+
+## Summarize Selected Features
+
+```{r}
+selected.features = cboost$getSelectedBaselearner()
+table(selected.features)
+```
+
+## Estimated Parameter
+
+```{r}
+params = cboost$getEstimatedCoef()
+str(params)
+```
+
+## Continue Training
+
+```{r}
+cboost$train(3000)
+```
+
+## Visualize Features
+
+```{r}
+cboost$plot("age_spline", iters = c(100, 500, 1000, 2000, 3000)) +
+  ggthemes::theme_tufte() + 
+  ggplot2::scale_color_brewer(palette = "Spectral")
+```
