### 2018-10-01 ed50dc2
fix i R installation for binder
