+++ binder/install.R
@@ -1,4 +1,4 @@
 install.packages("ggplot2")
-install.pacakges("ggthemes")
+install.packages("ggthemes")
 install.packages("mlbench")
 remotes::isntall_github("schalkdaniel/compboost")
