### 2018-10-01 0bd867b
update installal.R
