### 2018-10-01 0101ce8
update install.R
