+++ binder/install.R
@@ -2,4 +2,4 @@ install.packages("ggplot2")
 install.packages("ggthemes")
 install.packages("mlbench")
 install.packages("remotes")
-remotes::isntall_github("schalkdaniel/compboost")
+remotes::install_github("schalkdaniel/compboost")
