### 2018-10-02 a420a67
add Teststing section to Readme
