### 2018-10-09 #967 @arfon
@schalkdaniel  - it doesn't look like `https://doi.org/10.2402/compboost.0.1.0` resolves yet?
