### 2018-10-12 #967 @schalkdaniel
@arfon - I have uploaded a new version. The archive is: https://doi.org/10.5281/zenodo.1460435

The problem was that zenodo does not register custom dois. Sorry for that.
