### 2018-10-12 #967 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
