### 2018-10-31 #967 @schalkdaniel
Hi @arfon, is it possible to make a small change to the paper? I (stupidly) used a wrong latex symbol and didn't recognize it during the review process. Instead of `\geq` it should be `\gg`.

It would be really great if I could update the paper with the correct symbol.

Thanks for your efforts!
