### 2018-10-31 #967 @arfon
> It would be really great if I could update the paper with the correct symbol.

Sure thing. Please go ahead and push the fix to the `master` branch of your repository and we can recompile the paper.
