### 2018-10-31 #967 @schalkdaniel
Thanks! @arfon I have pushed the fix to the master. :-)
