### 2018-10-31 #967 @arfon
> Thanks! @arfon I have pushed the fix to the master. :-)

OK, I've updated the paper. This sometimes take a few hours to update on the live site because of caching.
